bug(map_env.py): a square of cells rotating in a clockwise direction would cause an infinite loop.
In the resolution mechanism, for moves that did not directly conflict with another agent over a cell, we would loop through and only resolve moves that were moving into an empty cell. However, in the case where every move is into a cell that currently has an agent in it, this will cause an infinite loop. We fix this by checking if any moves have been resolved, if none have we just let them all move