Change plotting so that only reward plots have a fixed bottom of 0.
This is done because reward tends to be deeply negative at the start, but quickly changes to a positive value. This makes the rest of the graph more legible, as the y-axis now covers a smaller area, so more detail can be seen, at the cost of hiding the negative values at the start.