Fix bug where MOA would use wrong observation: it would use the observation one timestep ahead.
This mix-up happened because the predictions are made post-hoc for computational efficiency: we don't have the actions a_t when evaluating the model at t.
The previous actions at t-1 were correctly passed to the MOA LSTM, but the observation at t would also be added, while we needed the observation at t-1.
This has been solved by passing the encoded MOA FC state to the state, so we can use it one step later.
Because the SCM depends on a certain size/order of the state, we had to increase the indexing in scm_model.py. This is brittle and should be changed, added TODO for this.